issue #74, modifying test up to date
